refactor: move tasks get usecases to a folder
