+++ src/main/factory/tasks/get/get-tasks-by-category.ts
@@ -1,4 +1,4 @@
-import { DbGetByTasks } from '@data/usecases/db/tasks/db-get-by-tasks'
+import { DbGetByTasks } from '@data/usecases/db/tasks/get/db-get-by-tasks'
 import { PgTasksRepository } from '@infra/usecases/db/postgres/tasks/pg-tasks-repository'
 import { GetTasksByCategoryController } from '@presentation/controller/tasks/get-by/get-tasks-by-category-controller'
 import { makeGetTasksByCategoryValidation } from './get-tasks-by-category-validation'
